Fix back-nav: YOLO status bar filter, identity threshold, selected tab detection
- Filter YOLO detections in status bar area (black border + iOS status bar
  height) to prevent calibration snapping to wrong icons
- Restore identity check threshold to 0.90 (regression from GUIClip removal)
  and use GUIClip visual similarity instead of EdgeIoU for is_same_page
- Add detect_selected_tab() to identify highlighted tab via focused LLM call,
  inject into planner context via _build_tab_context for correct tab navigation
- Fix numpy.float32 serialization error in icon_detector ONNX output
- Fix test_back_nav depth increment on failed tap (for→while)
- Strengthen planner Q2 prompt with ordered three-step tab identification
- Add xianyu_chat_back_btn eval case

+++ policy_expr/recon/icon_detector.py
@@ -148,15 +148,15 @@ class IconDetector:
         # Scale back to original image coordinates
         bboxes: list[IconBbox] = []
         for i in range(len(boxes_640)):
-            x1 = (boxes_640[i, 0] - pad_w) / ratio
-            y1 = (boxes_640[i, 1] - pad_h) / ratio
-            x2 = (boxes_640[i, 2] - pad_w) / ratio
-            y2 = (boxes_640[i, 3] - pad_h) / ratio
+            x1 = float((boxes_640[i, 0] - pad_w) / ratio)
+            y1 = float((boxes_640[i, 1] - pad_h) / ratio)
+            x2 = float((boxes_640[i, 2] - pad_w) / ratio)
+            y2 = float((boxes_640[i, 3] - pad_h) / ratio)
             # Clamp to image bounds
-            x1 = max(0, min(x1, iw))
-            y1 = max(0, min(y1, ih))
-            x2 = max(0, min(x2, iw))
-            y2 = max(0, min(y2, ih))
+            x1 = max(0.0, min(x1, iw))
+            y1 = max(0.0, min(y1, ih))
+            x2 = max(0.0, min(x2, iw))
+            y2 = max(0.0, min(y2, ih))
             bboxes.append(IconBbox(x1, y1, x2, y2, float(scores[i])))
 
         return bboxes
